refactor: move JSON Schema provider in its own package
